+++ src/main/java/org/example/controllers/LineController.java
@@ -1,36 +1,23 @@
 package org.example.controllers;
 
 import org.example.dto.LineDTO;
-import org.example.models.Line;
-import org.example.repositories.LineRepository;
+import org.example.services.LineService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/lines")
 public class LineController {
 
     @Autowired
-    private LineRepository lineRepository;
+    private LineService lineService;
 
     @GetMapping
     public List<LineDTO> getAllLines() {
-        return lineRepository
-                .findAll()
-                .stream()
-                .map(this::lineToDTO)
-                .collect(Collectors.toList());
-    }
-
-    private LineDTO lineToDTO(Line line) {
-        LineDTO lineDTO = new LineDTO();
-        lineDTO.setId(line.getId());
-        lineDTO.setName(line.getName());
-        return lineDTO;
+        return lineService.getAllLines();
     }
 }
