+++ src/main/resources/apllication.properties
@@ -3,6 +3,8 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
 
 spring.jpa.hibernate.ddl-auto=update
